requirements.txt
yaml~=0.2.5
pyyaml~=5.3.1
keras-gpu
pandas
opencv